requirements.txt
pywebview==4.4.1
jinja2==3.1.6
python-dotenv==1.2.1
paramiko==4.0.0